% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_beta.R
\name{b_beta}
\alias{b_beta}
\title{B-beta function of GMM model}
\usage{
b_beta(lambda, mu, sigma, beta = 0.001, rel.tol = 1e-12)
}
\arguments{
\item{lambda}{`numeric` - Vector of mixing proportions of each normal component}

\item{mu}{`numeric` - Vector of means of each normal component}

\item{sigma}{`numeric` - Vector of standard deviations of each normal component}

\item{beta}{`numeric(1)` - parameter corresponds to the beta-divergence induced likelihood}

\item{rel.tol}{relative accuracy requested.}
}
\value{
`numeric(1)`
}
\description{
\eqn{b_\beta(\theta) = \dfrac{1}{1 + \beta} \int f(x; \theta) ^ {1 + \beta} d x}
}
\examples{
set.seed(42)
lambda <- c(0.25, 0.75)
mu <- c(0, 4)
sigma <- c(1, 1)
b_beta(lambda, mu, sigma, beta=0.2)
b_beta(lambda, mu, sigma, beta=1e-3)
}
