% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_likelihood.R
\name{beta_likelihood}
\alias{beta_likelihood}
\title{Beta-Likelihood of Gaussian Mixture Model}
\usage{
beta_likelihood(x, lambda, mu, sigma, beta = 0.001, rel.tol = 1e-12)
}
\arguments{
\item{x}{`numeric(n)` - A numeric vector of observations}

\item{lambda}{`numeric` - Vector of mixing proportions of each normal component}

\item{mu}{`numeric` - Vector of means of each normal component}

\item{sigma}{`numeric` - Vector of standard deviations of each normal component}

\item{beta}{`numeric(1)` - parameter corresponds to the beta-divergence induced likelihood}

\item{rel.tol}{relative accuracy requested.}
}
\value{
`numeric(1)` - calculated beta-likelihood
}
\description{
\eqn{l_\beta(\theta) = \dfrac{1}{n \beta} \sum_{i=1}^n f(x_i; \beta)^{\beta} - b_\beta(\theta)}, where
\eqn{b_\beta(\theta) = \dfrac{1}{1 + \beta} \int f(x; \theta) ^ {1 + \beta} d x}
}
\examples{
set.seed(42)
lambda <- c(0.25, 0.75)
mu <- c(0, 4)
sigma <- c(1, 1)
x <- rnormix(n=100, lambda, mu, sigma)
beta_likelihood(x, lambda, mu, sigma, beta=0.2)
beta_likelihood(x, lambda, mu, sigma, beta=1e-3)
}
