% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_responsibility.R
\name{beta_responsibility}
\alias{beta_responsibility}
\title{Generalized (Beta) Responsibility}
\usage{
beta_responsibility(x, lambda, mu, sigma, beta)
}
\arguments{
\item{x}{`numeric(n)` A vector with n observations from a Gaussian Mixture Model.}

\item{lambda}{`numeric` - Vector of mixing proportions of each normal component}

\item{mu}{`numeric` - Vector of means of each normal component}

\item{sigma}{`numeric` - Vector of standard deviations of each normal component}

\item{beta}{`numeric(1)` - parameter corresponds to the beta-divergence induced likelihood.}
}
\value{
`matrix` A matrix with n rows and k columns representing the generalized responsibilities.
}
\description{
The generalized beta responsibility of a Gaussian Mixture Model (GMM) is \eqn{\gamma_{i, k}(\beta) = \dfrac{\lambda_k N(x_i \mid \mu_k, \Sigma_k)}{f(x_i; \theta)^{1 - \beta}}}. When `beta = 0`, \eqn{\gamma_{i, k}(0) = \Pr(Z_{k} = 1 \mid x_i)}.
}
\examples{
set.seed(42)
lambda <- c(0.25, 0.75)
mu <- c(0, 4)
sigma <- c(1, 1)
x <- rnormix(n=4, lambda, mu, sigma)
beta_responsibility(x, lambda, mu, sigma, beta=0)
beta_responsibility(x, lambda, mu, sigma, beta=0.2)
}
