% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnormix.R
\name{dnormix}
\alias{dnormix}
\title{Density of normal mixture}
\usage{
dnormix(x, lambda, mu, sigma, log = FALSE)
}
\arguments{
\item{x}{`numeric` - Vector of values at which to evaluate the density}

\item{lambda}{`numeric` - Vector of mixing proportions of each normal component}

\item{mu}{`numeric` - Vector of means of each normal component}

\item{sigma}{`numeric` - Vector of standard deviations of each normal component}

\item{log}{`logical` - Logical value indicating whether to return log density (TRUE) or density (FALSE)}
}
\value{
`numeric`; Vector of densities or log densities
}
\description{
Calculate probability density function of a mixture of univariate normal distributions.
}
\examples{
dnormix(-5:5, c(0.7, 0.3), c(0, 3), c(1, 1))

# visualize
p <- function(x) {dnormix(x, c(0.7, 0.3), c(0, 3), c(1, 1))}
lp <- function(x) {dnormix(x, c(0.7, 0.3), c(0, 3), c(1, 1), log=TRUE)}
par(mfrow=c(1, 2))
curve(p, -5, 5)
curve(lp, -5, 5)
par(mfrow=c(1, 1))

# supports fast numerical integration without underflow/overflow
integrate(p, -Inf, Inf) # equals 1
integrate(function(x){- p(x) * lp(x)}, -Inf, Inf) # entropy

}
