% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo_cvm_div.R
\name{loo_cvm_div}
\alias{loo_cvm_div}
\title{Leave-One-Out Cramer-von Mises Divergence}
\usage{
loo_cvm_div(x, lambda, mu, sigma, beta = 0.2, tol = 1e-05, maxiter = 1000)
}
\arguments{
\item{x}{`numeric(n)` - A numeric vector of observations}

\item{lambda}{`numeric` - Vector of mixing proportions of each normal component}

\item{mu}{`numeric` - Vector of means of each normal component}

\item{sigma}{`numeric` - Vector of standard deviations of each normal component}

\item{beta}{`numeric(1)` - parameter corresponds to the beta-divergence induced likelihood}

\item{maxiter}{Max iteration}
}
\description{
Leave-One-Out Cramer-von Mises Divergence
}
