% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lse.R
\name{lse}
\alias{lse}
\title{Log-Sum-Exp}
\usage{
lse(x)
}
\arguments{
\item{x}{numeric}
}
\value{
LSE
}
\description{
\eqn{LSE(x_1, \ldots, x_n) = \log(\exp(x_1) + \dots + \exp(x_n))}
}
\examples{
lse(c(-1001, -1002, -1003))
}
