% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnormix.R
\name{pnormix}
\alias{pnormix}
\title{Cumulative distribution function of normal mixture}
\usage{
pnormix(x, lambda, mu, sigma, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{x}{`numeric` Vector of values at which to evaluate the (log, complementary) cdf.}

\item{lambda}{`numeric` - Vector of mixing proportions of each normal component}

\item{mu}{`numeric` - Vector of means of each normal component}

\item{sigma}{`numeric` - Vector of standard deviations of each normal component}

\item{lower.tail}{`logical` - if TRUE (default), probabilities are `P[X ≤ x]` otherwise, `P[X > x]`.}

\item{log.p}{`logical`; if TRUE, probabilities p are given as `log(p)`.}
}
\value{
`numeric` Vector of (log) probabilities.
}
\description{
Cumulative distribution function of normal mixture
}
\examples{
pnormix(-5:5, c(0.7, 0.3), c(0, 3), c(1, 1))

}
