% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnormix.R
\name{rnormix}
\alias{rnormix}
\title{Generate random samples from normal mixture}
\usage{
rnormix(n, lambda, mu, sigma, method = c("standard", "montecarlo", "fix"))
}
\arguments{
\item{n}{`numeric(1)` - number of samples}

\item{lambda}{`numeric` - Vector of mixing proportions of each normal component}

\item{mu}{`numeric` - Vector of means of each normal component}

\item{sigma}{`numeric` - Vector of standard deviations of each normal component}

\item{method}{"standard" behaves like [rnorm]. "montecarlo" is more efficient
and recommended for Monte Carlo computation based on mixture samples; the
samples are ordered based on each component. "fixed" will use expected number
of samples in each component and produce a `data.frame`.}
}
\value{
`numeric` if "standard" or "montecarlo"; `data.frame` if "fixed"
}
\description{
Generate random samples from normal mixture
}
\examples{
set.seed(42)
rnormix(10, c(0.5, 0.5), c(-10, 10), c(1, 1), "standard")
rnormix(10, c(0.5, 0.5), c(-10, 10), c(1, 1), "montecarlo")
rnormix(10, c(0.5, 0.5), c(-10, 10), c(1, 1), "fix")

}
