% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid_normix_input.R
\name{valid_normix_input}
\alias{valid_normix_input}
\title{Check if inputs are valid for *normix()}
\usage{
valid_normix_input(lambda, mu, sigma)
}
\arguments{
\item{lambda}{`numeric` - Vector of mixing proportions of each normal component}

\item{mu}{`numeric` - Vector of means of each normal component}

\item{sigma}{`numeric` - Vector of standard deviations of each normal component}
}
\description{
Check if inputs are valid for *normix()
}
